docker_compose/app/requirements.txt
django==3.2
django-extensions==3.2.1
flake8==4.0
python-dotenv==0.21.0
django-split-settings==1.2.0
psycopg2-binary
django-debug-toolbar==3.4.0
pytest==7.2.0
gunicorn==20.1.0